Move check functions
- Moved check circulant functions to the tests.
- Make change to non-interactive plotting temporary.
- Added extra check for the transport direction.
