+++ tests/qttools/toeplitz/test_circulant.py
@@ -4,12 +4,99 @@
 
 import pytest
 
-from qttools import xp
-from qttools.toeplitz.circulant import (
-    check_nested_phi_circulant,
-    construct_circulant_cell,
-    expand_transverse,
-)
+from qttools import NDArray, xp
+from qttools.toeplitz.circulant import construct_circulant_cell, expand_transverse
+
+
+def check_phi_circulant(
+    a: NDArray,
+    sections: int,
+    phase: complex = 1.0,
+) -> bool:
+    """Check if a matrix is block phi circulant with the given number of sections.
+
+    Parameters
+    ----------
+    a : NDArray
+        The matrix to check.
+    sections : int
+        The number of sections in the block circulant structure.
+    phase : complex, optional
+        The phase factor to apply to the blocks, by default 1.0.
+
+    Returns
+    -------
+    bool
+        True if a is block circulant with the given number of sections, False otherwise.
+
+    """
+    if a.shape[-1] % sections != 0:
+        raise ValueError("The last dimension of a must be divisible by sections.")
+
+    if a.shape[-2] != a.shape[-1]:
+        raise ValueError(
+            "The second to last dimension of a must be equal to the last dimension of a."
+        )
+
+    block_size = a.shape[-1] // sections
+    # Take the first block-row (top n rows)
+    block_layer = a[..., :block_size, :]
+    blocks = xp.split(block_layer, sections, axis=-1)
+
+    for i in range(sections):
+        if i == 0:
+            shifted_blocks = blocks
+        else:
+            # Scale wrapped blocks by phase, leave remaining blocks unscaled
+            shifted_blocks = [phase * b for b in blocks[-i:]] + blocks[:-i]
+
+        expected = xp.concatenate(shifted_blocks, axis=-1)
+        actual = a[..., i * block_size : (i + 1) * block_size, :]
+
+        if not xp.allclose(actual, expected):
+            return False
+
+    return True
+
+
+def check_nested_phi_circulant(
+    a: NDArray,
+    sections: tuple[int, int],
+    phases: tuple[complex, complex] = (1.0, 1.0),
+) -> bool:
+    """Check if a matrix is block phi circulant in 2D with the given
+    number of sections.
+
+    Parameters
+    ----------
+    a : NDArray
+        The matrix to check.
+    sections : tuple[int, int]
+        The number of sections in the block circulant structure for each
+        dimension.
+    phases : tuple[complex, complex], optional
+        The phase factors to apply to the blocks for each dimension, by
+        default (1.0, 1.0).
+
+    Returns
+    -------
+    bool
+        True if a is block circulant in 2D with the given number of
+        sections, False otherwise.
+
+    """
+    if not check_phi_circulant(a, sections[1], phases[1]):
+        return False
+
+    for i in range(sections[1]):
+        for j in range(sections[1]):
+            block_size = a.shape[-1] // sections[1]
+            slice_i = slice(i * block_size, (i + 1) * block_size)
+            slice_j = slice(j * block_size, (j + 1) * block_size)
+            if not check_phi_circulant(a[slice_i, slice_j], sections[0], phases[0]):
+                return False
+
+    return True
 
 
 def _generate_dict(
